output: discard metrics with empty data series
When a metrics context is about to be flushed, it may pass through one or more
processors before reaching the output plugin. Processing can result in a valid
metrics context with no time series data, which is useless for output plugins.
This change detects that case, skips the flush, and returns FLB_OK to the engine.

Signed-off-by: Eduardo Silva <eduardo@chronosphere.io>